+++ src/main/java/io/github/JavaProject_DauphineM1_2019/anonymizeData/AnonymiseAndWriteData.java
@@ -1,0 +1,5 @@
+package io.github.JavaProject_DauphineM1_2019.anonymizeData;
+
+public class AnonymiseAndWriteData {
+
+}
