+++ src/test/java/io/github/JavaProject_DauphineM1_2019/checkingData/RulesTest.java
@@ -4,6 +4,8 @@ import static org.junit.jupiter.api.Assertions.*;
 
 import org.junit.jupiter.api.Test;
 
+import io.github.JavaProject_DauphineM1_2019.Rules.Rules;
+
 class RulesTest {
 	
 	Rules r = new Rules();
